+++ golibs/commander/commander.go
@@ -17,8 +17,11 @@ In this example, a command with the name of `hello` is created
 that will print `Hello world!` to output. One question you may
 have is: What is the `sinks` parameter?
 
-The `sinks` parameter is a table with 3 keys: `in`, `out`,
-and `err`. All of them are a @Sink.
+The `sinks` parameter is a table with 3 keys: `input`, `out`, and `err`.
+There is an `in` alias to `input`, but it requires using the string accessor syntax (`sinks['in']`)
+as `in` is also a Lua keyword, so `input` is preferred for use.
+All of them are a @Sink.
+In the future, `sinks.in` will be removed.
 
 - `in` is the standard input.
 You may use the read functions on this sink to get input from the user.
@@ -40,11 +43,13 @@ import (
 type Commander struct{
 	Events *bait.Bait
 	Loader packagelib.Loader
+	Commands map[string]*rt.Closure
 }
 
-func New(rtm *rt.Runtime) Commander {
-	c := Commander{
+func New(rtm *rt.Runtime) *Commander {
+	c := &Commander{
 		Events: bait.New(rtm),
+		Commands: make(map[string]*rt.Closure),
 	}
 	c.Loader = packagelib.Loader{
 		Load: c.loaderFunc,
@@ -58,6 +63,7 @@ func (c *Commander) loaderFunc(rtm *rt.Runtime) (rt.Value, func()) {
 	exports := map[string]util.LuaExport{
 		"register": util.LuaExport{c.cregister, 2, false},
 		"deregister": util.LuaExport{c.cderegister, 1, false},
+		"registry": util.LuaExport{c.cregistry, 0, false},
 	}
 	mod := rt.NewTable()
 	util.SetExports(rtm, mod, exports)
@@ -88,7 +94,7 @@ func (c *Commander) cregister(t *rt.Thread, ct *rt.GoCont) (rt.Cont, error) {
 		return nil, err
 	}
 
-	c.Events.Emit("commandRegister", cmdName, cmd)
+	c.Commands[cmdName] = cmd
 
 	return ct.Next(), err
 }
@@ -105,7 +111,23 @@ func (c *Commander) cderegister(t *rt.Thread, ct *rt.GoCont) (rt.Cont, error) {
 		return nil, err
 	}
 
-	c.Events.Emit("commandDeregister", cmdName)
+	delete(c.Commands, cmdName)
 
 	return ct.Next(), err
+}
+
+// registry() -> table
+// Returns all registered commanders. Returns a list of tables with the following keys:
+// - `exec`: The function used to run the commander. Commanders require args and sinks to be passed.
+// #returns table
+func (c *Commander) cregistry(t *rt.Thread, ct *rt.GoCont) (rt.Cont, error) {
+	registryLua := rt.NewTable()
+	for cmdName, cmd := range c.Commands {
+		cmdTbl := rt.NewTable()
+		cmdTbl.Set(rt.StringValue("exec"), rt.FunctionValue(cmd))
+
+		registryLua.Set(rt.StringValue(cmdName), rt.TableValue(cmdTbl))
+	}
+
+	return ct.PushingNext1(t.Runtime, rt.TableValue(registryLua)), nil
 }