+++ rl.go
@@ -70,11 +70,8 @@ func newLineReader(prompt string, noHist bool) *lineReader {
 		hooks.Emit("hilbish.vimAction", actionStr, args)
 	}
 	rl.HintText = func(line []rune, pos int) []rune {
-		if hinter == nil {
-			return []rune{}
-		}
-
-		retVal, err := rt.Call1(l.MainThread(), rt.FunctionValue(hinter),
+		hinter := hshMod.Get(rt.StringValue("hinter"))
+		retVal, err := rt.Call1(l.MainThread(), hinter,
 		rt.StringValue(string(line)), rt.IntValue(int64(pos)))
 		if err != nil {
 			fmt.Println(err)
@@ -89,10 +86,8 @@ func newLineReader(prompt string, noHist bool) *lineReader {
 		return []rune(hintText)
 	}
 	rl.SyntaxHighlighter = func(line []rune) string {
-		if highlighter == nil {
-			return string(line)
-		}
-		retVal, err := rt.Call1(l.MainThread(), rt.FunctionValue(highlighter),
+		highlighter := hshMod.Get(rt.StringValue("highlighter"))
+		retVal, err := rt.Call1(l.MainThread(), highlighter,
 		rt.StringValue(string(line)))
 		if err != nil {
 			fmt.Println(err)